backend/requirements.txt
# Backend Dependencies

# Web Framework
fastapi==0.135.1
uvicorn==0.41.0
starlette==0.52.1

# Database
SQLAlchemy==2.0.48
psycopg2-binary==2.9.11

# Authentication & Security
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.0.1
cryptography==46.0.5
python-dotenv==1.2.2

# Validation
email-validator==2.3.0
pydantic==2.12.5

# Computer Vision & AI
opencv-python==4.13.0.92
tensorflow==2.21.0
keras==3.13.2
ultralytics==8.4.26

# Data Processing
numpy==2.4.3
pandas==2.3.3
Pillow==12.1.1

# Windows-specific (Media)
pygrabber==0.2

# Utilities
requests==2.32.5
h11==0.16.0
